+++ src/containerapp/azext_containerapp/containerapp_auth_decorator.py
@@ -2,13 +2,12 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-from typing import Any, Dict
+# pylint: disable=line-too-long
 
 from azure.cli.command_modules.containerapp.containerapp_auth_decorator import ContainerAppAuthDecorator
 from azure.cli.command_modules.containerapp._utils import safe_set
 
 from ._constants import BLOB_STORAGE_TOKEN_STORE_SECRET_SETTING_NAME
-from knack.prompting import prompt_y_n
 from azure.cli.core.azclierror import ArgumentUsageError
 
 
@@ -32,10 +31,6 @@ class ContainerAppPreviewAuthDecorator(ContainerAppAuthDecorator):
             raise ArgumentUsageError('Usage Error: only blob storage token store is supported. --sas-url-secret and --sas-url-secret-name should provide exactly one when token store is enabled')
         if self.get_argument_sas_url_secret() is not None and self.get_argument_sas_url_secret_name() is not None:
             raise ArgumentUsageError('Usage Error: --sas-url-secret and --sas-url-secret-name cannot both be set')
-        if self.get_argument_sas_url_secret() is not None and not self.get_argument_yes():
-            msg = 'Configuring --sas-url-secret will add a secret to the containerapp. Are you sure you want to continue?'
-            if not prompt_y_n(msg, default="n"):
-                raise ArgumentUsageError('Usage Error: --sas-url-secret cannot be used without agreeing to add secret to the containerapp.')
 
         sas_url_setting_name = BLOB_STORAGE_TOKEN_STORE_SECRET_SETTING_NAME
         if self.get_argument_sas_url_secret_name() is not None: