+++ src/monitor-control-service/azext_amcs/aaz/latest/monitor/data_collection/rule/association/_create.py
@@ -22,9 +22,9 @@ class Create(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2022-06-01",
+        "version": "2023-03-11",
         "resources": [
-            ["mgmt-plane", "/{resourceuri}/providers/microsoft.insights/datacollectionruleassociations/{}", "2022-06-01"],
+            ["mgmt-plane", "/{resourceuri}/providers/microsoft.insights/datacollectionruleassociations/{}", "2023-03-11"],
         ]
     }
 
@@ -133,7 +133,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2022-06-01",
+                    "api-version", "2023-03-11",
                     required=True,
                 ),
             }
@@ -225,6 +225,10 @@ class Create(AAZCommand):
             metadata = cls._schema_on_200_201.properties.metadata
             metadata.provisioned_by = AAZStrType(
                 serialized_name="provisionedBy",
+                flags={"read_only": True},
+            )
+            metadata.provisioned_by_immutable_id = AAZStrType(
+                serialized_name="provisionedByImmutableId",
                 flags={"read_only": True},
             )
             metadata.provisioned_by_resource_id = AAZStrType(