+++ src/connectedvmware/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+1.1.0
+++++++
+* Prompt credentials if not provided for Guest Agent.
+
 1.0.1
 ++++++
 * Fix bug in `az connectedvmware create` command when RG of machine is different from RG of vCenter.